V1.0
TODO:吸收时的动量计算，Join的计算效率，轨迹预测,相机推近

+++ Assets/SpaceWorld/SimpleForce/ForceSender/SimapleForceSender_C.cs
@@ -2,8 +2,8 @@ using Unity.Entities;
 using Unity.Mathematics;
 
 public struct SimapleForceSender_C : IComponentData {
-    public float time;
+    public double time;
     public Entity to;//受力点所属
-    public float3 value; //均匀力大小
+    public double3 value; //均匀力大小
     public ForceType type;//施力类型
 }
