+++ Assets/SpaceWorld/SimpleForce/Force_C.cs
@@ -2,8 +2,8 @@ using Unity.Entities;
 using Unity.Mathematics;
 
 public struct Force_C : IComponentData {
-    public float3 value;
-    public float time;
+    public double3 value;
+    public double time;
     public ForceType type;
     public Entity from; //施力点
     public Entity to; //受力点
