+++ Assets/SpaceWorld/SimpleForce/Momentum/Momentum_C.cs
@@ -3,9 +3,9 @@ using Unity.Mathematics;
 
 public struct Momentum_C : IComponentData {
     //质量
-    public float mass;
+    public double mass;
     //速度
-    public float3 speed;
+    public double3 speed;
     ////动量增加给的mover
     //public Mover_C mover;
     // mover所在实体
